+++ src/common/utils/deployTips.ts
@@ -118,8 +118,8 @@ Example of element to delete:
       expressionRegex: [/In field: template - no EmailTemplate named (.*) found/gm],
       tip: `Lightning EmailTemplates must also be imported with metadatas.
 ${c.cyan(
-  "If this type of error is displayed in a deployment with --check, you may ignore it and validate the PR anyway (it may not happen when the deployment will be really performed and split in steps, including the one importing EmailTemplate records)"
-)}
+        "If this type of error is displayed in a deployment with --check, you may ignore it and validate the PR anyway (it may not happen when the deployment will be really performed and split in steps, including the one importing EmailTemplate records)"
+      )}
 - Create a file scripts/data/EmailTemplates/export.json:
 {
     "objects": [
@@ -133,11 +133,29 @@ ${c.cyan(
 - Run sfdx hardis:work:save`,
     },
     {
+      name: "empty-item",
+      label: "Empty source items",
+      expressionString: [
+        "Required field is missing: sharingOwnerRules",
+        "Required field is missing: standardValue",
+        "Required field is missing: valueTranslation"
+      ],
+      tip: `You probably retrieved empty items, that must not be included within the SFDX project
+To remove them, please run sfdx:hardis:project:clean:emptyitems`,
+    },
+    {
       name: "field-must-not-be-required",
       label: "Formula picklist field issue",
       expressionRegex: [/Field:(.*) must not be Required/gm],
       tip: `You probably made read only a field that was required before.
 Find the field in the layout source XML, then replace Required by Readonly`,
+    },
+    {
+      name: "field-not-available-for-element",
+      label: "Field not available for element",
+      expressionRegex: [/Field (.*) is not available for/gm],
+      tip: `You probably changed the type of a field.
+Find the field in the source XML, and remove the section using it`,
     },
     {
       name: "formula-picklist-issue",
@@ -156,8 +174,8 @@ More details at https://help.salesforce.com/articleView?id=sf.tips_on_building_f
     {
       name: "missing-field-middle-name",
       label: "Missing field MiddleName",
-      expressionString: ["field MiddleName"],
-      tip: `Quotes must be activated in the target org.
+      expressionString: ["field MiddleName", "Variable does not exist: MiddleName"],
+      tip: `MiddleNames must be activated in the target org.
 - Help: https://help.salesforce.com/articleView?id=000332623&type=1&mode=1
 - Scratch org setting: 
 "nameSettings": {
@@ -168,24 +186,34 @@ More details at https://help.salesforce.com/articleView?id=sf.tips_on_building_f
       name: "missing-field-suffix",
       label: "Missing field Suffix",
       expressionString: ["field Suffix"],
-      tip: `Quotes must be activated in the target org.
+      tip: `Suffix must be activated in the target org.
 - Help: https://help.salesforce.com/articleView?id=000332623&type=1&mode=1
 - Scratch org setting: 
 "nameSettings": {
   "enableNameSuffix": true
 },`,
     },
-
     {
       name: "missing-field-synced-quote-id",
       label: "Missing field SyncedQuoteId",
-      expressionString: ["field SyncedQuoteId"],
+      expressionString: ["field SyncedQuoteId",
+        "Error  force-app/main/default/objects/Quote/Quote.object-meta.xml",
+        "Error  force-app/main/default/objects/Opportunity/fields/SyncedQuoteId.field-meta.xml"],
       tip: `Quotes must be activated in the target org.
 - Help: https://help.salesforce.com/articleView?id=sf.quotes_enable.htm&type=5
 - Scratch org setting: 
 "quoteSettings": {
   "enableQuote": true
 }`,
+    },
+    {
+      name: "missing-feature-account-contact-relation",
+      label: "Missing feature ContactToMultipleAccounts",
+      expressionString: ["no CustomObject named AccountContactRelation found"],
+      tip: `Contacts to multiple accounts be activated in the target org.
+- Help: https://help.salesforce.com/articleView?id=sf.shared_contacts_set_up.htm&type=5
+- Scratch org setting: 
+"features": ["ContactsToMultipleAccounts"]`,
     },
     {
       name: "missing-feature-chatter-collaboration-groups",
@@ -235,6 +263,17 @@ More details at https://help.salesforce.com/articleView?id=sf.tips_on_building_f
 "languageSettings": {
   "enableTranslationWorkbench":  true,
   "enableEndUserLanguages": true
+}`,
+    },
+    {
+      name: "missing-feature-opportinity",
+      label: "Missing feature Opportunity Teams",
+      expressionString: ["OpportunityTeam"],
+      tip: `Opportunity Teams must be activated in the target org.
+- Org: Setup -> Opportunity Team Settings -> Enable Team Selling
+- Scratch org:
+"opportunitySettings": {
+  "enableOpportunityTeam": true
 }`,
     },
     {